Index des fonctions

Documentation de toute les fonctions du package EDO

EDO.d_fun_vdp

EDO.fun_vdp

EDO.ode_euler

EDO.ode_euler_pasfixe

EDO.ode_gauss

EDO.ode_gauss_newton

EDO.ode_gauss_pf

EDO.ode_heun

EDO.ode_rk41

EDO.ode_rk42

EDO.ode_runge

EDO.d_fun_vdp — Method

dérivée du deuxieme membre de L'équation différentielle considérée est l'équation de Van der Pol

Inputs

- t = l'instant t  
- y = la valeur de y à cet instant

Outputs

dypoint = $\left[\begin{array}{cc} 0 & 1 \\ -2 y_1(t)y_2(t)-1 & 1-y_1(t)^2 \end{array} \right]$

EDO.fun_vdp — Method

Deuxieme membre de L'équation différentielle considérée est l'équation de Van der Pol

Inputs

t = l'instant t

y = la valeur de y à cet instant

Outputs

ypoint = $\left\{ \begin{aligned}\begin{array}{c} \dot{y}_{1}(t)= &y_{2}(t) \\ \dot{y}_{2}(t)= &(1-y_{1}^{2}(t)) y_{2}(t)-y_{1}(t) \end{array} \end{aligned}\right.$

EDO.ode_euler — Method

Explicit Euler method

Description

    Numerical integration of the Cauchy's problem
    x_point(t) = phi(t,x(t))
    x(t_0) = x_0

Usage

- T, X = ode_euler(phi,t0tf,y0,N)

Inputs

phi    - function     : second member of the ode whith the interface 
                      xpoint = phi(t, x)
                          t    - real     : time,
                          x = vector of R^n with the same dimension of x0
t0tf - real(2)      : intial and final time  [t0,tf]
x0   - real(n)      : initial point
N    - integer      : number of steps (>1)

Outputs

T    - real(N+1,1)  : vector of times
X    - real(N+1,n)  : Matrix of solution, The line i of [T Y] contains ti and x_i

EDO.ode_euler_pasfixe — Method

Programme d'integration par le schema d'Euler à pas constant

Usage

- T,Y,nphie,ifail=ode_euler_pasfixe(phi,t0tf,y0,N)

Input parameters

- phi = second member ypoint=phi[t,y]
- t0tf = [t0,tf]
- y0 = initial point
- N = number of step

Output parameters

- T = vector of time
- Y = Matrix of solution, The line i of [T Y] contains ti & y[ti]

EDO.ode_gauss — Method

Numerical integration by Gauss method of order 4

Usage

- T,Y,nphie,ifail=ode_gauss(phi,t0tf,y0,options)

Input parameters

- phi = second member ypoint=phi[t,y]
- t0tf = [t0,tf]
- y0 = initial point
- options[1] = N = number of step
- options[2] = fpitermax = maximum number of iterations for the fixed point
- options[3] = fpeps = epsilon for the test of progress in the fixed point

Output parameters

- T = vector of time
- Y = Matrix of solution The line i of [T Y] contains ti & y[ti]
- ifail[i] = 1 = computation successful for the fixed point on [t_i,t_[i+1]] 
- ifail[i] = -1 = computation failed for the fixed point on [t_i,t_[i+1]]: maximum number of iteration is attained in the fixed point
- nphie = number of evaluation of phi

EDO.ode_gauss_newton — Method

Numerical integration by Gauss method of order 4,The solution on the non linear equation is computed by Newton method

Usage

- T,Y,nphie,ndphie,ifail=ode_gauss_newton(phi,dphi,t0tf,y0,options)

Input parameters

- phi = second member ypoint=phi(t,y)
- t0tf = [t0,tf]
- y0 = initial point
- options[1] = N = number of step
- options[2] = fpitermax = maximum number of iterations for the fixed point
- options[3] = fpeps = epsilon for the test of progress in the fixed point

Output parameters

- T = vector of times
- Y = Matrix of solution, The line i of [T Y] contains ti and y(ti)
- ifail[i] = 1 = computation successful for the fixed point on [t_i,t_{i+1}] 
- ifail[i] = -1 = computation failed for the fixed point on [t_i,t_{i+1}]: maximum number of iteration is attained in the fixed point
- nphie = number of evaluation of phi

EDO.ode_gauss_pf — Method

Numerical integration by Gauss method of order 4

Usage

- [T,Y,nphie,ifail,KK]=ode_gauss_pf(phi,t0tf,y0,options)

Input parameters

- phi = second member ypoint=phi(t,y)
- t0tf = [t0,tf]
- y0 = initial point
- options(1) = N = number of step
- options(2) = fpitermax = maximum number of iterations for the fixed point
- options(3) = fpeps = epsilon for the test of progress in the fixed point

Output parameters

- T = vector of time
- Y = Matrix of solution, The line i of [T Y] contains ti and y(ti)
- ifail(i) = 1 = computation successful for the fixed point on [t_i,t_{i+1}] 
- ifail(i) = -1 = computation failed for the fixed point on [t_i,t_{i+1}]: maximum number of iteration is attained in the fixed point
- nphie = number of evaluation of phi

EDO.ode_heun — Method

Heun method

Description

Numerical integration of the Cauchy's problem
x_point(t) = f(t,x(t))
x(t_0) = x_0

Usage

T, X = ode_heun(f,t0tf,y0,N)

Inputs

f    - function     : second member of the ode whith the interface 
                      xpoint = f(t, x)
                          t    - real     : time,
                          x = vector of R^n with the same dimension of x0
t0tf - real(2)      : intial and final time  [t0,tf]
x0   - real(n)      : initial point
N    - integer      : number of steps (>1)

Outputs

T    - real(N+1,1)  : vector of times
X    - real(N+1,n)  : Matrix of solution
The line i of [T Y] contains ti and x_i

EDO.ode_rk41 — Method

Heun method

Description

Numerical integration of the CauchX[i-1,:]'s problem
x_point(t) = phi(t,x(t))
x(t_0) = x_0

Usage

T, X = ode_rk4(phi,t0tf,X[i-1,:]0,N)

Inputs

f    - function     : second member of the ode whith the interface 
                      xpoint = phi(t, x)
                          t    - real     : time,
                          x = vector of R^n with the same dimension of x0
t0tf - real(2)      : intial and final time  [t0,tf]
x0   - real(n)      : initial point
N    - integer      : number of steps (>1)

Outputs

T    - real(N+1,1)  : vector of times
X    - real(N+1,n)  : Matrix of solution, The line i of [T Y] contains ti and x_i

EDO.ode_rk42 — Method

Heun method

Description

Numerical integration of the CauchX[i-1,:]'s problem
x_point(t) = phi(t,x(t))
x(t_0) = x_0

Usage

T, X = ode_rk4(phi,t0tf,X[i-1,:]0,N)

Inputs

f    - function     : second member of the ode whith the interface 
                      xpoint = phi(t, x)
                          t    - real     : time,
                          x = vector of R^n with the same dimension of x0
t0tf - real(2)      : intial and final time  [t0,tf]
x0   - real(n)      : initial point
N    - integer      : number of steps (>1)

Outputs

T    - real(N+1,1)  : vector of times
X    - real(N+1,n)  : Matrix of solution, The line i of [T Y] contains ti and x_i

EDO.ode_runge — Method

runge method

Description

Numerical integration of the Cauchy's problem
x_point(t) = phi(t,x(t))
x(t_0) = x_0

Usage

- T, X = ode_runge(phi,t0tf,y0,N)

Inputs

f    - function     : second member of the ode whith the interface 
                      xpoint = phi(t, x)
                          t    - real     : time,
                          x = vector of R^n with the same dimension of x0
t0tf - real(2)      : intial and final time  [t0,tf]
x0   - real(n)      : initial point
N    - integer      : number of steps (>1)

Outputs

T    - real(N+1,1)  : vector of times
X    - real(N+1,n)  : Matrix of solution, The line i of [T Y] contains ti and x_i